Filling pattern of code from both continuous and discrete examples. Generated files are in visu-compo/product

+++ visu-compo-meta-model/src/main/java/utils/FileOperation.java
@@ -3,11 +3,25 @@ package utils;
 import java.io.*;
 import java.net.URL;
 import java.net.URLConnection;
+import java.nio.file.Files;
+import java.nio.file.LinkOption;
+import java.nio.file.Path;
 
 /**
  * Created by Ivan Logre on 24/06/2014.
  */
 public class FileOperation {
+    /*
+     * Read a local text file, extract and return the content as a String
+     */
+    public static String getStringFromFile(String dataSource) {
+        //the extraction is not the same if it's a remote file or a local one.
+        if (dataSource.contains("http://") || dataSource.contains("localhost"))
+           return FileOperation.getStringFromRemoteFile(dataSource);
+        else
+            return FileOperation.getStringFromLocalFile(dataSource);
+    }
+
 /*
  * Read a local text file, extract and return the content as a String
  */
@@ -56,18 +70,22 @@ public class FileOperation {
     *       the given path is valid
     */
     //TODO test limits (existing file or not, append etc)
-    public static void setDataFileFromObject(Object dataToPrint, String filePath, String encoding) {
+    public static void fillFileFromObject(Object dataToPrint, String filePath, String encoding) {
         try {
+            File f = new File(filePath);
+            if (!f.getParentFile().exists()){
+                throw new IOException("---> Please check that your path is valid (all mentioned directories have to exist)");
+            }
             Writer writer = new BufferedWriter(new OutputStreamWriter(new FileOutputStream(filePath), encoding));
             writer.write(dataToPrint.toString());
             writer.close();
-        } catch (IOException e) {
-            System.err.println("setDataFileFromObject caused this. Untested yet.");
-            e.printStackTrace();
-        }
+            } catch (IOException e) {
+                System.err.println("---> Method FileOperation.fillFileFromObject() caused this. Untested yet.");
+                e.printStackTrace();
+            }
     }
 
-    public static void setDataFileFromObject(Object dataToPrint, String filePath) {
-         setDataFileFromObject(dataToPrint, filePath, "utf-8");
+    public static void fillFileFromObject(Object dataToPrint, String filePath) {
+         fillFileFromObject(dataToPrint, filePath, "utf-8");
     }
 }
